Build fix for Xcode 3.1: Set InfoPlist.strings as UTF-8 in the project.
With it set to UTF-16, iconv was trying to convert the UTF-8 data from UTF-16 to UTF-16, resulting in an “incomplete character or shift sequence error”. Additionally, viewing it in Xcode showed what appeared to be Asian characters.

With it set to UTF-8, iconv now converts the UTF-8 data from UTF-8 to UTF-16, and this succeeds.